refactor: extract clusters/kubeconfig.rs tests into sibling file
Move the ~380 LOC test module out of kubeconfig.rs into
kubeconfig_tests.rs, included via #[cfg(test)] #[path = ...] mod tests.

Shrinks kubeconfig.rs from 521 to 146 LOC (production code only)
while keeping the same test module path and all 14 tests passing.
